Add support for control plane operator deployed from an OCP payload
This commit adds support for deploying the control-plane-operator from an OCP
payload. To enable this, the hypershift-operator has been updated to inspect the
release specified by a HostedCluster and, if a `hypershift` image is present in
the release metadata, that image will be used for the control-plane-operator.
Otherwise all the prior lookup behavior remains the same, so the code should be
compatible with all existing use cases and should work transparently when an OCP
payload containing hypershift is specified.

This commit also fixes the token minter, availability prober, and the socks
proxy deployments to ensure they are deployed by default from the hypershift
image instead of the control-plane-operator's image, as those components are
versioned independently of the control plane. A new control-plane-operator
`--token-minter-image` flag was added to support this.

Finally, this commit reduces the control plane image to include only the
control-plane-operator, which is the single component which will be shipped in
the OCP payload.