Fix route53 wildcard entry cleanup
For public zones, the wildcard entry name was wrong, it has the cluster
name as a component and that can not be derived from the zone name.

For private zones, it was needlessly complicated, the wildcard entry is
simply '*.apps.' + zoneNameWithTrailingDotStripped.
zoneNameWithTrailingDotStripped is the same as clusterName + "." +
baseDomain, but by directly using the zone name it will work even if the
basedomain argument is incorrect.